dctpro/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
Faker==36.1.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
psycopg==3.2.5
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.1
